+++ openpharmacophore/io/__init__.py
@@ -4,5 +4,3 @@ from .pharmacophore_mol2 import load_mol2_pharmacophoric_points, mol2_file_info
 from .moe import ph4_string, pharmacophoric_points_from_ph4_file
 from .mol2 import load_mol2_ligands, mol2_iterator
 from .mol_file import mol_file_to_list, mol_file_iterator
-from .mol_io import MolIO
-from .traj_io import TrajIO
