--- openpharmacophore/io/mol_io.py
@@ -1,5 +1,0 @@
-from openpharmacophore.io.file_io import AbstractIO
-
-
-class MolIO(AbstractIO):
-    pass
