+++ openpharmacophore/molecular_systems/binding_site.py
@@ -1,0 +1,4 @@
+
+
+class BindingSite:
+    pass
