+++ openpharmacophore/molecular_systems/ligand.py
@@ -1,8 +1,11 @@
 
-
 class Ligand:
     pass
 
 
 class LigandSet:
+    pass
+
+
+def create_ligand_set(filename: str):
     pass
