+++ openpharmacophore/molecular_systems/topology.py
@@ -1,0 +1,10 @@
+import mdtraj
+
+
+class Topology:
+    """ Wrapper for mdtraj's topology object."""
+    pass
+
+
+def create_topology(file_name: str):
+    pass
