+++ openpharmacophore/pharmacophore/ligand_based/rdp.py
@@ -1,4 +1,4 @@
-from openpharmacophore.chem_feats.chem_feats import feature_centroids
+from openpharmacophore.molecular_systems.chem_feats import feature_centroids
 from openpharmacophore.utils.maths import points_distance
 from openpharmacophore import PharmacophoricPoint, Pharmacophore
 import numpy as np
