+++ tests/unit/chem_feats/test_chem_feats.py
@@ -1,4 +1,4 @@
-from openpharmacophore.chem_feats.chem_feats import feature_indices, feature_centroids
+from openpharmacophore.molecular_systems.chem_feats import feature_indices, feature_centroids
 from rdkit import Chem
 import numpy as np
