+++ tests/unit/pharmacophore/ligand_receptor/test_pl_complex.py
@@ -2,7 +2,7 @@ import mdtraj as mdt
 import numpy as np
 import openpharmacophore.pl_complex.exceptions as exc
 from openpharmacophore import PLComplex
-from openpharmacophore.chem_feats.chem_feats import smarts_ligand
+from openpharmacophore.molecular_systems.chem_feats import smarts_ligand
 import pyunitwizard as puw
 import pytest
 from copy import deepcopy
